Add initial "framework workflow"
Begin work towards providing a common workflow or framework
for creating Nagios plugins.

This commit includes code from the current `atc0005/check-cert`
project (`v0.1.3` as of this work) and also provides some
initial example code to help get a feel for the changes.

More work and additional polish is planned.

refs GH-6