Adjust formatting of Nagios Performance Data
Update format for emitted performance data metrics to use single space
as separator instead of newline.

While the multi-line format used previously is recognized and
supported by Nagios, Icinga did not interpret the performance data in
the same way. Testing confirms that this updated format works with
both monitoring platforms.

This commit also updates doc comments and ensures that a trailing
newline is retained as the final line in generated output to satisfy
the Nagios plugin output format.

- refs atc0005/go-nagios#81
- refs atc0005/check-vmware#321