Add initial plugin output test
I hesitate to call this a true End to End (e2e) test, but this
initial test configures an ExitState value as client code in the
wild currently does and then asserts that the output matches
the contents of a carefully crafted testdata file.

NOTE:

This test adds a dependency on the popular
github.com/google/go-cmp testing package. Because future tests
are planned which will also use this package, adding the
dependency (only for tests) is deemed to be worth the "cost" of
adding an external dependency (again, just for tests).