Add check_cert plugin perfdata success test case
Add test case for lifetime metrics which uses `%` UoM.

NOTE:

As of this writing the plugin does not emit threshold values
for this metric so our test case does not specify thresholds
either. If thresholds are provided in the future for this
metric we should update this test case OR add an additional
test case which specifies thresholds.