Add support for (internal) debug logging output
OVERVIEW

The changes provided by this commit were initially intended to provide
a minimal abstraction to support toggling the size of plugin output
off/on as needed for troubleshooting. From there the scope widened as
further iteration showed that it could be useful to add debug logging
support (internally for now) to the library as a whole.

To support toggling on/off specific debug log activity, debug log
output is currently split into two categories:

- debug logging of actions
- debug logging of plugin output size

The intent of these changes is to provide a way to easily toggle
on/off the internal workings of this library. If client code does not
enable debug logging, the result should be this library working just
as it did before.

The current implementation is subject to change as real world testing
is applied and feedback collected.

CHANGES

- allow toggling debug logging output on/off
- allow toggling specific debug logging output on/off
  - e.g., allow all debug messages except for general plugin activity
  - e.g., allow only debug messages for plugin output size calculations
- support setting custom debug logging output target
  - e.g., redirect debug log messages to a target file instead of
    stderr (the default debug log target)
- add test coverage for new behavior

REFERENCES

- refs GH-264
- refs GH-271