Unescape encoded Ascii85 payload before decoding
Brief testing shows that when the encoded Ascii85 payload passes
through the Nagios XI (JSON) API that the backslash character is
escaped. Instead of handling this client-side we extend the
decoding process to automatically handle unescaping that
pattern. This also provides a "hook" for further unescaping if
other monitoring systems apply similar escaping of the encoded
payload.

Also, fix various typos in doc comments not caught during
previous polishing passes.

refs GH-251